update structure and fix bugs
