+++ test/test_ontology_sdk.py
@@ -10,7 +10,7 @@ from ontology.core.program import ProgramBuilder
 from ontology.crypto.signature_scheme import SignatureScheme
 from ontology.ont_sdk import OntologySdk
 
-sdk = OntologySdk()
+sdk = OntologySdk.get_instance()
 rpc_address = "http://polaris3.ont.io:20336"
 sdk.rpc.set_address(rpc_address)
 private_key = "523c5fcf74823831756f0bcb3634234f10b3beb1c05595058534577752ad2d9f"
@@ -50,5 +50,3 @@ class TestOntologySdk(TestCase):
         self.assertEqual("03036c12be3726eb283d078dff481175e96224f0b0c632c7a37e10eb40fe6be889", a[0].hex())
         self.assertEqual("020f9ce29ede5f0e271b67e61b2480dccc98c3aabad095c604ef9ab1d92a475c0a", a[1].hex())
         self.assertEqual("035384561673e76c7e3003e705e4aa7aee67714c8b68d62dd1fb3221f48c5d3da0", a[2].hex())
-
-
